requirements.txt
streamlit
fitz
groq
pandas
apify-client
python-dotenv
PyMuPDF
langchain
euriai